가게 메뉴 조회 기능

+++ src/main/java/com/example/smartorder/dto/OrderMenuDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class OrderMenuDto {
 
 	String menuName;
-	double menuPrice;
+	long menuPrice;
 	int menuCount;
 
 }
